requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
coloredlogs==15.0.1
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
distro==1.6.0
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
dominate==2.6.0
fire==0.4.0
flatbuffers==1.12
fonttools==4.26.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
html5lib==1.1
humanfriendly==9.2
idna==3.2
imageio==2.9.0
img2pdf==0.4.1
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.2
networkx==2.6.2
numpy==1.19.5
oauthlib==3.1.1
ocrmypdf==12.3.2
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
parso==0.8.2
pdf2docx==0.5.2
pdfminer==20191125
pdfminer.six==20201018
pdftotree==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==2.16.1
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.10.0
PyMuPDF==1.18.16
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
PyWavelets==1.1.1
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
selectivesearch==0.4
six==1.15.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tabula-py==2.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2021.8.8
tqdm==4.62.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
Wand==0.6.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
